add files write

+++ remapping-driver.h
@@ -3,6 +3,6 @@
 
 static int init_remapping(void);
 static void exit_remapping(void);
-static void write_something(const char *, int*);
+static void write_something(const char *,u64*);
 static void dump_buffer(void);
 #endif
